+++ shared/src/androidMain/kotlin/com/jinkyu/tv/presentation/login/AndroidLoginViewModel.kt
@@ -2,8 +2,6 @@ package com.jinkyu.tv.presentation.login
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jinkyu.tv.presentation.LoginViewModel
-import com.jinkyu.tv.presentation.RegisterViewModel
 
 class AndroidLoginViewModel : ViewModel() {
