+++ shared/src/androidMain/kotlin/com/jinkyu/tv/presentation/login/LoginRoute.kt
@@ -1,9 +1,6 @@
 package com.jinkyu.tv.presentation.login
 
 import androidx.compose.runtime.Composable
-import com.jinkyu.tv.LoginScreen
-import com.jinkyu.tv.RegisterScreen
-import com.jinkyu.tv.navigation.Login
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
