+++ shared/src/androidMain/kotlin/com/jinkyu/tv/presentation/register/AndroidRegisterViewModel.kt
@@ -2,7 +2,6 @@ package com.jinkyu.tv.presentation.register
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jinkyu.tv.presentation.RegisterViewModel
 
 class AndroidRegisterViewModel : ViewModel() {
