+++ shared/src/androidMain/kotlin/com/jinkyu/tv/presentation/register/RegisterRoute.kt
@@ -1,16 +1,17 @@
 package com.jinkyu.tv.presentation.register
 
 import androidx.compose.runtime.Composable
-import com.jinkyu.tv.RegisterScreen
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
 fun RegisterRoute(
+    navigateMain: () -> Unit,
     navigateLogin: () -> Unit,
     viewModel: AndroidRegisterViewModel = koinViewModel()
 ) {
     RegisterScreen(
         navigateLogin = navigateLogin,
+        navigateMain = navigateMain,
         viewModel = viewModel.viewModel
     )
 }
