+++ shared/src/commonMain/kotlin/com/jinkyu/tv/presentation/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.jinkyu.tv.presentation
+package com.jinkyu.tv.presentation.login
 
 import com.jinkyu.tv.domain.user.UserInput
 import kotlinx.coroutines.CoroutineScope
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 
-class RegisterViewModel(
+class LoginViewModel(
     private val coroutineScope: CoroutineScope? = null
 ) {
     private val viewModelScope = coroutineScope ?: CoroutineScope(Dispatchers.Main)
@@ -18,14 +18,11 @@ class RegisterViewModel(
     private val _email: MutableStateFlow<String> = MutableStateFlow<String>("")
     val email: StateFlow<String> = _email.asStateFlow()
 
-    private val _nickName: MutableStateFlow<String> = MutableStateFlow<String>("")
-    val nickName: StateFlow<String> = _nickName.asStateFlow()
-
     private val _password: MutableStateFlow<String> = MutableStateFlow<String>("")
     val password: StateFlow<String> = _password.asStateFlow()
 
-    val registerEnable: StateFlow<Boolean> = combine(_email, _nickName, _password) { email, nickName, password ->
-        email.isNotBlank() && nickName.isNotBlank() && password.isNotBlank()
+    val loginEnable: StateFlow<Boolean> = combine(_email, _password) { email, password ->
+        email.isNotBlank() && password.isNotBlank()
     }.stateIn(
         scope = viewModelScope,
         started = SharingStarted.WhileSubscribed(5000),
@@ -37,15 +34,15 @@ class RegisterViewModel(
 
     fun onUserInput(type: UserInput, input: String) {
         when (type) {
-            UserInput.NICKNAME -> _nickName.value = input
             UserInput.EMAIL -> _email.value = input
             UserInput.PASSWORD -> _password.value = input
+            else -> {}
         }
     }
 
     fun onSignUpClicked() {}
 
-    fun onBackButtonClicked() {}
+    fun onLoginClicked() {}
 
     fun onRememberMeClicked(enable: Boolean) {
         _rememberMe.value = enable
