+++ shared/src/commonMain/kotlin/com/jinkyu/tv/presentation/register/RegisterNavigationAction.kt
@@ -1,0 +1,6 @@
+package com.jinkyu.tv.presentation.register
+
+sealed class RegisterNavigationAction {
+    data object NavigateToMain : RegisterNavigationAction()
+    data object NavigateLogin : RegisterNavigationAction()
+}
