+++ shared/src/commonMain/kotlin/com/jinkyu/tv/presentation/register/RegisterViewModel.kt
@@ -1,28 +1,38 @@
-package com.jinkyu.tv.presentation
+package com.jinkyu.tv.presentation.register
 
 import com.jinkyu.tv.domain.user.UserInput
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.launch
 
-class LoginViewModel(
+class RegisterViewModel(
     private val coroutineScope: CoroutineScope? = null
 ) {
     private val viewModelScope = coroutineScope ?: CoroutineScope(Dispatchers.Main)
 
+    private val _navigationAction: MutableSharedFlow<RegisterNavigationAction> = MutableSharedFlow<RegisterNavigationAction>()
+    val navigationAction: SharedFlow<RegisterNavigationAction> = _navigationAction.asSharedFlow()
+
     private val _email: MutableStateFlow<String> = MutableStateFlow<String>("")
     val email: StateFlow<String> = _email.asStateFlow()
+
+    private val _nickName: MutableStateFlow<String> = MutableStateFlow<String>("")
+    val nickName: StateFlow<String> = _nickName.asStateFlow()
 
     private val _password: MutableStateFlow<String> = MutableStateFlow<String>("")
     val password: StateFlow<String> = _password.asStateFlow()
 
-    val loginEnable: StateFlow<Boolean> = combine(_email, _password) { email, password ->
-        email.isNotBlank() && password.isNotBlank()
+    val registerEnable: StateFlow<Boolean> = combine(_email, _nickName, _password) { email, nickName, password ->
+        email.isNotBlank() && nickName.isNotBlank() && password.isNotBlank()
     }.stateIn(
         scope = viewModelScope,
         started = SharingStarted.WhileSubscribed(5000),
@@ -34,15 +44,19 @@ class LoginViewModel(
 
     fun onUserInput(type: UserInput, input: String) {
         when (type) {
+            UserInput.NICKNAME -> _nickName.value = input
             UserInput.EMAIL -> _email.value = input
             UserInput.PASSWORD -> _password.value = input
-            else -> {}
         }
     }
 
     fun onSignUpClicked() {}
 
-    fun onLoginClicked() {}
+    fun onBackButtonClicked() {
+        viewModelScope.launch {
+            _navigationAction.emit(RegisterNavigationAction.NavigateLogin)
+        }
+    }
 
     fun onRememberMeClicked(enable: Boolean) {
         _rememberMe.value = enable
